refactor: 공통 코드 분리
