+++ src/lib/utils.tsx
@@ -1,0 +1,8 @@
+import { usePathname } from "next/navigation";
+
+export function usePath(){
+    const pathname = usePathname();
+    const id = pathname.split('/').pop();
+
+    return id;
+}
